validation

+++ BooksStore/BookStoreClient/BookClient.cs
@@ -17,8 +17,13 @@ namespace BookStoreClient
             _client = client;
         }
 
-        public async Task<Book> CreateBook(Book book)
+        public async Task<Book> CreateBookAsync(Book book)
         {
+            if (book == null)
+            {
+                return new Book();
+            }
+
             HttpContent httpContent = new StringContent(JsonConvert.SerializeObject(book), Encoding.UTF8);
             var response = await _client.PostAsync("/api/book", httpContent);
 
@@ -33,8 +38,13 @@ namespace BookStoreClient
             }
         }
 
-        public async Task<bool> DeleteBook(int id)
+        public async Task<bool> DeleteBookAsync(int id)
         {
+            if (id == 0)
+            {
+                return false;
+            }
+
             var response = await _client.DeleteAsync($"/api/book/{id}");
 
             try
@@ -48,10 +58,14 @@ namespace BookStoreClient
             }
         }
 
-        public async Task<Book> EditBook(Book book)
+        public async Task<Book> EditBookAsync(Book book)
         {
-            HttpContent httpContent = new StringContent(JsonConvert.SerializeObject(book), Encoding.UTF8);
+            if (book == null)
+            {
+                return new Book();
+            }
 
+            HttpContent httpContent = new StringContent(JsonConvert.SerializeObject(book), Encoding.UTF8);
             var response = await _client.PutAsync($"/api/book/{book.BookId}", httpContent);
 
             try
@@ -66,10 +80,11 @@ namespace BookStoreClient
         }
 
         public async Task<IEnumerable<Book>> GetAllBooksAsync()
-        {            
+        {      
+            var response = await _client.GetAsync("/api/book");
+
             try
             {
-                var response = await _client.GetAsync("/api/book");
                 response.EnsureSuccessStatusCode();
                 return JsonConvert.DeserializeObject<IEnumerable<Book>>(await response.Content.ReadAsStringAsync());
             }
@@ -79,8 +94,13 @@ namespace BookStoreClient
             }
         }
 
-        public async Task<Book> GetBookById(int id)
+        public async Task<Book> GetBookByIdAsync(int id)
         {
+            if (id == 0)
+            {
+                return new Book();
+            }
+
             var response = await _client.GetAsync($"/api/book/{id}");
 
             try
